indexer: add lib.rs to allow using crate in another binary
